+++ src/lib.rs
@@ -9,7 +9,8 @@ use tokio::{net::UdpSocket, runtime, task::JoinSet};
 mod buffer;
 mod protocol;
 mod query;
-mod new_cache;
+mod cache;
+mod min_heap;
 
 const DOWNSTREAM_IP_ADDR: Ipv4Addr = Ipv4Addr::new(0, 0, 0, 0);
 const DOWNSTREAM_PORT: u16 = 1234;
